+++ test/e2e/framework/providers/gce/firewall_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package framework
+package gce
 
 import "testing"
