+++ test/e2e/network/scale/localrun/ingress_scale.go
@@ -33,6 +33,8 @@ import (
 	gcecloud "k8s.io/kubernetes/pkg/cloudprovider/providers/gce"
 
 	"k8s.io/kubernetes/test/e2e/framework"
+	"k8s.io/kubernetes/test/e2e/framework/ingress"
+	"k8s.io/kubernetes/test/e2e/framework/providers/gce"
 	"k8s.io/kubernetes/test/e2e/network/scale"
 )
 
@@ -117,7 +119,7 @@ func main() {
 		glog.Errorf("Error building GCE provider: %v", err)
 		os.Exit(1)
 	}
-	cloudConfig.Provider = gceCloud
+	cloudConfig.Provider = gce.NewProvider(gceCloud)
 
 	testSuccessFlag := true
 	defer func() {
@@ -150,7 +152,7 @@ func main() {
 
 	// Setting up a localized scale test framework.
 	f := scale.NewIngressScaleFramework(cs, ns.Name, cloudConfig)
-	f.Logger = &framework.GLogger{}
+	f.Logger = &ingress.GLogger{}
 	// Customizing scale test.
 	f.EnableTLS = enableTLS
 	f.OutputFile = outputFile
