+++ test/e2e/storage/regional_pd.go
@@ -37,6 +37,7 @@ import (
 	kubeletapis "k8s.io/kubernetes/pkg/kubelet/apis"
 	"k8s.io/kubernetes/pkg/volume/util"
 	"k8s.io/kubernetes/test/e2e/framework"
+	"k8s.io/kubernetes/test/e2e/framework/providers/gce"
 	"k8s.io/kubernetes/test/e2e/storage/utils"
 	imageutils "k8s.io/kubernetes/test/utils/image"
 )
@@ -210,7 +211,7 @@ func testZonalFailover(c clientset.Interface, ns string) {
 		waitStatus <- waitForStatefulSetReplicasNotReady(statefulSet.Name, ns, c)
 	}()
 
-	cloud, err := framework.GetGCECloud()
+	cloud, err := gce.GetGCECloud()
 	if err != nil {
 		Expect(err).NotTo(HaveOccurred())
 	}
